requirements.txt
# annotated-types==0.6.0
# anyio==4.3.0
APScheduler==3.10.4
# azure-core==1.30.1
azure-storage-blob==12.19.1
amazon-textract-caller==0.2.4
amazon-textract-prettyprinter==0.1.10
azure-ai-formrecognizer==3.3.3
beautifulsoup4==4.12.3
# blinker==1.7.0
boto3==1.34.62
botocore==1.34.62
# certifi==2024.2.2
# cffi==1.16.0
# chardet==5.2.0
# charset-normalizer==3.3.2
# click==8.1.7
# cryptography==42.0.5
# distro==1.9.0
# dnspython==2.6.1
# et-xmlfile==1.1.0
# exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Executor==1.0.0
gunicorn==21.2.0
# h11==0.14.0
# httpcore==1.0.4
# httpx==0.27.0
# idna==3.6
# isodate==0.6.1
# itsdangerous==2.1.2
# Jinja2==3.1.3
# jmespath==1.0.1
# lxml==5.1.0
# MarkupSafe==2.1.5
Markdown==3.7
Markdown2docx==0.1.0
natsort==8.4.0
numpy==1.26.4
openai==1.14.0
openpyxl==3.1.2
pandas==2.2.1
# pdfminer.six==20231228
pdfplumber==0.11.0
# pillow==10.2.0
# pycparser==2.21
# pydantic==2.6.4
# pydantic_core==2.16.3
pymongo==4.6.2
PyMuPDF==1.23.26
PyPDF2==3.0.1
# PyMuPDFb==1.23.22
# pypdfium2==4.28.0
# pyspnego==0.10.2
# python-dateutil==2.9.0.post0
pytz==2024.1
python-docx==1.1.2
Pillow==9.5.0
reportlab==4.1.0
requests==2.31.0
# requests-ntlm==1.2.0
# requests-toolbelt==1.0.0
# s3transfer==0.10.0
# SharePlum==0.5.1
# six==1.16.0
# sniffio==1.3.1
# soupsieve==2.5
tabula-py==2.9.0
# tqdm==4.66.2
# typing_extensions==4.10.0
# tzdata==2024.1
# tzlocal==5.2
ultralytics==8.2.90
# urllib3==2.0.7
# Werkzeug==3.0.1